pluggable errorDecoder on conjureRuntime
