+++ .palantir/revapi.yml
@@ -274,3 +274,106 @@ acceptedBreaks:
       new: "method com.palantir.dialogue.clients.DialogueClients.StickyChannelFactory2\
         \ com.palantir.dialogue.clients.DialogueClients.ReloadingFactory::getStickyChannels2(java.lang.String)"
       justification: "New @Beta API"
+  "3.68.0":
+    com.palantir.dialogue:dialogue-serde:
+    - code: "java.class.kindChanged"
+      old: "enum com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      new: "interface com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.class.noLongerFinal"
+      old: "enum com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      new: "interface com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.class.noLongerImplementsInterface"
+      old: "enum com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      new: "interface com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.class.noLongerInheritsFromClass"
+      old: "enum com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      new: "interface com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.class.nowAbstract"
+      old: "enum com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      new: "interface com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.field.removed"
+      old: "field com.palantir.conjure.java.dialogue.serde.ErrorDecoder.INSTANCE"
+      justification: "Used by annotations"
+    - code: "java.method.nowAbstract"
+      old: "method boolean com.palantir.conjure.java.dialogue.serde.ErrorDecoder::isError(com.palantir.dialogue.Response)"
+      new: "method boolean com.palantir.conjure.java.dialogue.serde.ErrorDecoder::isError(com.palantir.dialogue.Response)"
+      justification: "Used by annotations"
+    - code: "java.method.nowAbstract"
+      old: "method java.lang.RuntimeException com.palantir.conjure.java.dialogue.serde.ErrorDecoder::decode(com.palantir.dialogue.Response)"
+      new: "method java.lang.RuntimeException com.palantir.conjure.java.dialogue.serde.ErrorDecoder::decode(com.palantir.dialogue.Response)"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method <T extends java.lang.Enum<T extends java.lang.Enum<T>>> T java.lang.Enum<E\
+        \ extends java.lang.Enum<E extends java.lang.Enum<E>>>::valueOf(java.lang.Class<T>,\
+        \ java.lang.String) @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method boolean java.lang.Object::equals(java.lang.Object) @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method com.palantir.conjure.java.dialogue.serde.ErrorDecoder com.palantir.conjure.java.dialogue.serde.ErrorDecoder::valueOf(java.lang.String)"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method com.palantir.conjure.java.dialogue.serde.ErrorDecoder[] com.palantir.conjure.java.dialogue.serde.ErrorDecoder::values()"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method int java.lang.Enum<E extends java.lang.Enum<E extends java.lang.Enum<E>>>::compareTo(E)\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method int java.lang.Enum<E extends java.lang.Enum<E extends java.lang.Enum<E>>>::ordinal()\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method int java.lang.Object::hashCode() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method java.lang.Class<?> java.lang.Object::getClass() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method java.lang.Class<E> java.lang.Enum<E extends java.lang.Enum<E extends\
+        \ java.lang.Enum<E>>>::getDeclaringClass() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method java.lang.Object java.lang.Object::clone() throws java.lang.CloneNotSupportedException\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method java.lang.String java.lang.Enum<E extends java.lang.Enum<E extends\
+        \ java.lang.Enum<E>>>::name() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method java.lang.String java.lang.Object::toString() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method java.util.Optional<java.lang.Enum.EnumDesc<E>> java.lang.Enum<E\
+        \ extends java.lang.Enum<E extends java.lang.Enum<E>>>::describeConstable()\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method void java.lang.Object::finalize() throws java.lang.Throwable @\
+        \ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method void java.lang.Object::notify() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method void java.lang.Object::notifyAll() @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method void java.lang.Object::wait() throws java.lang.InterruptedException\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method void java.lang.Object::wait(long) throws java.lang.InterruptedException\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
+    - code: "java.method.removed"
+      old: "method void java.lang.Object::wait(long, int) throws java.lang.InterruptedException\
+        \ @ com.palantir.conjure.java.dialogue.serde.ErrorDecoder"
+      justification: "Used by annotations"
