+++ dialogue-serde/src/test/java/com/palantir/conjure/java/dialogue/serde/BinaryEncodingTest.java
@@ -34,7 +34,7 @@ public class BinaryEncodingTest {
         TestResponse response = new TestResponse().code(200).contentType("application/octet-stream");
         BodySerDe serializers = new ConjureBodySerDe(
                 ImmutableList.of(WeightedEncoding.of(new ConjureBodySerDeTest.StubEncoding("application/json"))),
-                ErrorDecoder.INSTANCE,
+                ConjureErrorDecoder.INSTANCE,
                 Encodings.emptyContainerDeserializer(),
                 DefaultConjureRuntime.DEFAULT_SERDE_CACHE_SPEC);
         InputStream deserialized = serializers.inputStreamDeserializer().deserialize(response);
@@ -58,7 +58,7 @@ public class BinaryEncodingTest {
         TestResponse response = new TestResponse().code(200).contentType("application/octet-stream");
         BodySerDe serializers = new ConjureBodySerDe(
                 ImmutableList.of(WeightedEncoding.of(new ConjureBodySerDeTest.StubEncoding("application/json"))),
-                ErrorDecoder.INSTANCE,
+                ConjureErrorDecoder.INSTANCE,
                 Encodings.emptyContainerDeserializer(),
                 DefaultConjureRuntime.DEFAULT_SERDE_CACHE_SPEC);
         Optional<InputStream> maybe =
