+++ dialogue-serde/src/test/java/com/palantir/conjure/java/dialogue/serde/DefaultClientsTest.java
@@ -80,7 +80,7 @@ public final class DefaultClientsTest {
     private Response response = new TestResponse();
     private BodySerDe bodySerde = new ConjureBodySerDe(
             DefaultConjureRuntime.DEFAULT_ENCODINGS,
-            ErrorDecoder.INSTANCE,
+            ConjureErrorDecoder.INSTANCE,
             Encodings.emptyContainerDeserializer(),
             DefaultConjureRuntime.DEFAULT_SERDE_CACHE_SPEC);
     private final SettableFuture<Response> responseFuture = SettableFuture.create();
